requirements.txt
arxiv>=1.4.2
httpx>=0.27.0
requests>=2.31.0
beautifulsoup4>=4.12.3
Pillow>=10.2.0  # Phase 9: 图片验证
pdf2image==1.16.3
openai>=1.45.0
redis>=5.0.8
tenacity>=8.3.0
python-dotenv>=1.0.1
pyyaml>=6.0.2
huggingface_hub>=0.24.0
feedparser>=6.0.10
pytest>=8.3.2
pytest-asyncio>=0.24.0
pydantic>=1.10.15
black>=24.8.0
ruff>=0.6.3
flask>=3.0.0
gunicorn>=21.2.0

# Phase 8: PDF解析
scipdf-parser==0.52  # 学术论文PDF解析 (基于GROBID)
lxml>=5.0.0  # XML解析器（优化PDF解析性能，消除BeautifulSoup警告）
